Extract ReporterScope
